Fix and move `test_multiple_consolidations_*` tests to sanity block tests
