+++ tests/core/pyspec/eth2spec/test/electra/sanity/blocks/__init__.py
@@ -1,1 +1,2 @@
+from .test_consolidation import *  # noqa: F401 F403
 from .test_deposit_transition import *  # noqa: F401 F403
